Add multi-cmgr mode (#17)
* Add multi-cmgr mode

* Link to fork in role readme

* Use fork URL in service templates

* Fix suffixed db path templating

* Fix loop variable output registration

* Fix service restart notify conditional groupings

* Remove service active status checks

I don't think it's possible for these to work with multi mode, because
we'd need to loop until every service in the registered result variable
is active.

I'm not sure how useful these are anyway, because they cause the play to
fail if e.g. the Docker servers are temporarily unreachable for some
reason, despite cmgrd/cmgr-artifact-server being successfully installed.
Some manual administration like restarting services is to be expected,
especially since cmgrd relies on an external host (the Docker engine
server) to start successfully.